refactor(mcp): move tool unit tests into their tool files
Each tool's unit tests now live in its own tools/<tool>.rs under a
#[cfg(test)] mod tests, next to the code they exercise. The four insta
golden snapshots move to src/tools/snapshots/ to match the new module
paths (bodies unchanged). The cross-cutting resources test stays in
tools.rs; shared test helpers (handler, src, fixtures) are pub(crate) at
the tools module root. No behavior change.
